+++ Cliente/app/src/main/java/com/example/entities/User.java
@@ -2,7 +2,7 @@ package com.example.entities;
 
 public class User {
     private String fullName;
-    private int age;
+    private Double age;
     private String email;
     private String address;
     private String type;
@@ -11,7 +11,7 @@ public class User {
 
     }
 
-    public User(String fullName, int age, String email, String address) {
+    public User(String fullName, Double age, String email, String address) {
         this.fullName = fullName;
         this.age = age;
         this.email = email;
@@ -36,11 +36,11 @@ public class User {
         this.fullName = fullName;
     }
 
-    public int getAge() {
+    public Double getAge() {
         return age;
     }
 
-    public void setAge(int age) {
+    public void setAge(Double age) {
         this.age = age;
     }
